+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [2-beta8] - 2025-02-25
 - update deps (metamorph.ml, Clay)
+- added Clay's main CLI to the uberjar build
 
 ## [2-beta7] - 2025-02-20
 - updated deps (tablecloth, kindly, tableplot, same/ish, clay)